added cahrge to convection, added the species, densities for the thagunna example, added the specie for the electron energy, added the border for the continuity equation

+++ inc/specie.hpp
@@ -17,7 +17,7 @@ class Specie{
 
 		double get_mass();
 
-		double get_charge();
+		int get_charge();
 
 		double get_qm_ratio();
